+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 ### Changed
 - Updated leapsecond information according to Bulletin C 55 (2018-01-09).
+- `6d6mseed` now mentions the creation of a logfile or an auxfile in the log.
 
 ### Fixed
 - Fixed a segfault in `6d6mseed` when output pattern contains no dirname.